+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog of lizard-graph
 0.1 (unreleased)
 ----------------
 
+- Initial working line graph, using timeseries, nens-graph and
+  lizard-fewsnorm.
+
 - Initial views and urls.py.
 
 - Initial library skeleton created by nensskel.  [Jack Ha]